Update the packing of all remaining classes that still use 32 bit reference    packing to use 64 bit packing. Reading is backwards compatible with old data    using the packing data version flags. Consequently data written with this    update can only be read by this version.
git-svn-id: http://svn.cern.ch/guest/lhcb/LHCb/trunk@184975 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 8540fbac81c3c9a2d9d85bf0815efa28adf1ddaf

+++ LHCb/Event/EventPacker/src/component/ReadPackedDst.cpp
@@ -187,7 +187,7 @@ StatusCode ReadPackedDst::execute() {
       put( caloHypos, name + m_postfix );
       processLinks( caloHypos, version );
       getFromBlob<LHCb::PackedCaloHypo> ( caloHypos->hypos() , blobs );
-      getFromBlob<int>                  ( caloHypos->refs()  , blobs );
+      getFromBlob<long long>            ( caloHypos->refs()  , blobs );
 
     } else if ( LHCb::CLID_PackedProtoParticles  == classID ) {
 
@@ -204,7 +204,7 @@ StatusCode ReadPackedDst::execute() {
       put( recVertices, name + m_postfix );
       processLinks( recVertices, version );
       getFromBlob<LHCb::PackedRecVertex> ( recVertices->vertices() , blobs );
-      getFromBlob<int>                   ( recVertices->refs()    , blobs );
+      getFromBlob<long long>             ( recVertices->refs()    , blobs );
       getFromBlob<std::pair<int,int> >   ( recVertices->extras() , blobs );
 
     } else if ( LHCb::CLID_PackedTwoProngVertices     == classID ) {
